# docs/swagger.yaml
info:
  contact: {}
paths:
  /event/create:
    post:
      consumes:
      - multipart/form-data
      description: 增加事件信息
      operationId: /event/create
      parameters:
      - in: formData
        name: name
        type: string
      - in: formData
        name: status
        type: string
      - in: formData
        name: type
        type: string
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            type: string
      summary: 增加事件
      tags:
      - 事件信息接口
  /event/delete:
    delete:
      consumes:
      - application/json
      description: 删除单个事件
      operationId: /event/delete
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: string
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            type: string
      summary: 删除事件
      tags:
      - 事件信息接口
  /event/events:
    get:
      consumes:
      - application/json
      description: 获取事件信息
      operationId: /event/events
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            type: string
      summary: 事件信息
      tags:
      - 事件信息接口
  /event/patch:
    patch:
      consumes:
      - multipart/form-data
      description: 修改事件信息
      operationId: /event/patch
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: string
      - in: formData
        name: name
        type: string
      - in: formData
        name: status
        type: string
      - in: formData
        name: type
        type: string
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            type: string
      summary: 修改事件
      tags:
      - 事件信息接口
  /login:
    post:
      consumes:
      - multipart/form-data
      description: 登录接口
      operationId: /user/Login
      parameters:
      - in: formData
        name: password
        type: string
      - in: formData
        name: username
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            type: string
      summary: 登录
      tags:
      - 用户信息接口
  /plan/create:
    post:
      consumes:
      - multipart/form-data
      description: 增加计划信息
      operationId: /plan/create
      parameters:
      - in: formData
        name: end
        type: string
      - in: formData
        name: start
        type: string
      - in: formData
        name: title
        type: string
      - in: formData
        name: user
        type: integer
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            type: string
      summary: 增加计划
      tags:
      - 计划信息接口
  /project/create:
    post:
      consumes:
      - multipart/form-data
      description: 增加项目信息
      operationId: /project/create
      parameters:
      - in: formData
        name: name
        type: string
      - in: formData
        name: status
        type: string
      - in: formData
        name: type
        type: string
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            type: string
      summary: 增加项目
      tags:
      - 项目信息接口
  /project/delete:
    delete:
      consumes:
      - application/json
      description: 删除单个项目
      operationId: /project/delete
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: string
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            type: string
      summary: 删除项目
      tags:
      - 项目信息接口
  /project/patch:
    patch:
      consumes:
      - multipart/form-data
      description: 修改项目信息
      operationId: /project/patch
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: string
      - in: formData
        name: name
        type: string
      - in: formData
        name: status
        type: string
      - in: formData
        name: type
        type: string
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            type: string
      summary: 修改项目
      tags:
      - 项目信息接口
  /project/projects:
    get:
      consumes:
      - application/json
      description: 获取项目信息
      operationId: /project/projects
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            type: string
      summary: 项目信息
      tags:
      - 项目信息接口
  /register:
    post:
      consumes:
      - multipart/form-data
      description: 增加用户信息
      operationId: /user/register
      parameters:
      - in: formData
        name: password
        type: string
      - in: formData
        name: username
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            type: string
      summary: 注册
      tags:
      - 用户信息接口
  /user/create:
    post:
      consumes:
      - multipart/form-data
      description: 增加用户信息
      operationId: /user/create
      parameters:
      - in: formData
        name: password
        type: string
      - in: formData
        name: username
        type: string
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            type: string
      summary: 增加用户
      tags:
      - 用户信息接口
  /user/delete:
    delete:
      consumes:
      - application/json
      description: 删除单个用户
      operationId: /user/delete
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: string
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            type: string
      summary: 删除用户
      tags:
      - 用户信息接口
  /user/patch:
    patch:
      consumes:
      - multipart/form-data
      description: 修改用户信息
      operationId: /user/patch
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: string
      - in: formData
        name: password
        type: string
      - in: formData
        name: username
        type: string
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            type: string
      summary: 修改用户
      tags:
      - 用户信息接口
  /user/users:
    get:
      consumes:
      - application/json
      description: 获取用户信息
      operationId: /user/index
      parameters:
      - description: Info
        in: query
        name: Info
        required: true
        type: string
      - description: PageSize
        in: query
        name: PageSize
        required: true
        type: string
      - description: PageNo
        in: query
        name: PageNo
        required: true
        type: string
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            type: string
      summary: 用户信息
      tags:
      - 用户信息接口
swagger: "2.0"